moves PersonalTabContent and MemberSubscriptionTypeBlock into default folder
