+++ lib/pages/home/default/homeWidget.dart
@@ -15,7 +15,7 @@ import 'package:readr_app/models/OnBoardingPosition.dart';
 import 'package:readr_app/models/section.dart';
 import 'package:readr_app/pages/tabContent/listening/listeningTabContent.dart';
 import 'package:readr_app/pages/tabContent/news/tabContent.dart';
-import 'package:readr_app/pages/tabContent/personal/personalTabContent.dart';
+import 'package:readr_app/pages/tabContent/personal/default/personalTabContent.dart';
 import 'package:readr_app/widgets/newsMarquee/newsMarquee.dart';
 import 'package:readr_app/widgets/popupRoute/easyPopup.dart';
 import 'package:readr_app/widgets/popupRoute/sectionDropDownMenu.dart';
